Move to lib/api, better type checking, use Promise.all()
